ARM64: dts: tissot: Fix the wrong fg-iterm-ma value
fg-iterm-ma stores the charge current at which fuel gauge starts
scaling the SoC to 100%. We had it set to 3000mA which is
ridiculously high and makes our SoC increase too fast without
reflecting the actual battery capacity.
Let's set it to 300mA which is the proper value for our fuel gauge.

Signed-off-by: Khusika Dhamar Gusti <mail@khusika.com>